Fix foreign key type mismatch in fellow subscriptions migrations
Change fellow_id from unsignedBigInteger to unsignedInteger to match
the existing fellows table int primary key. Also add hasTable guard
to cardiothoracic migration for idempotent re-runs.

+++ database/migrations/2026_02_26_164038_create_fellow_subscriptions_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('fellow_subscriptions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('fellow_id');
+            $table->unsignedInteger('fellow_id');
             $table->string('year', 10);
             $table->enum('status', ['Paid', 'Unpaid', 'Partial', 'Waived'])->default('Unpaid');
             $table->decimal('amount_due', 10, 2)->default(0);
